requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.7
colorama==0.4.6
Cython==3.0.11
Django==4.2.16
django-cors-headers==4.5.0
djangorestframework==3.15.2
graphql-core==2.3.2
importlib_metadata==8.5.0
packaging==24.1
pillow==11.0.0
pip-tools==7.4.1
promise==2.3
psycopg2-binary==2.9.4
pyproject_hooks==1.2.0
python-dotenv==1.0.1
Rx==1.6.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.20.2
